+++ CHANGELOG.md
@@ -17,6 +17,8 @@ This is the changelog page for Lemon Tools. _Most,_ but not all changes will be 
   - String UPPERCASE
   - String esreveR
 
+## Changes
+
 # 2.1.1
 
 ## What's new